Enforce correct usage of gl::Version
In preparation for the implementation update,
ensured that gl::Version struct is always
initialized with 8-bit values.

Bug: angleproject:408843436
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6458758
Commit-Queue: Alexey Knyazev <lexa.knyazev@gmail.com>

+++ src/libANGLE/Version.h
@@ -9,13 +9,15 @@
 #ifndef LIBANGLE_VERSION_H_
 #define LIBANGLE_VERSION_H_
 
+#include <cstdint>
+
 namespace gl
 {
 
 struct Version
 {
     constexpr Version();
-    constexpr Version(unsigned int major, unsigned int minor);
+    constexpr Version(uint8_t major, uint8_t minor);
 
     unsigned int major;
     unsigned int minor;
